updated vendor and customer resource

+++ app/Http/Controllers/CustomerController.php
@@ -9,7 +9,7 @@ use App\Http\Requests\Customer\UpdateRequest;
 use App\Http\Resources\CustomerResource;
 use App\Http\Resources\CustomerResourceCollection;
 use App\Repositories\Contracts\CustomerRepository;
-use Illuminate\Http\Request;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class CustomerController extends Controller
 {
@@ -32,9 +32,12 @@ class CustomerController extends Controller
      *
      * @param IndexRequest $request
      * @return CustomerResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [Customer::class, $request->get('propertyId')]);
+
         $customers = $this->customerRepository->findBy($request->all());
 
         return new CustomerResourceCollection($customers);
@@ -45,9 +48,12 @@ class CustomerController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return CustomerResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [Customer::class, $request->get('propertyId')]);
+
         $customer = $this->customerRepository->save($request->all());
 
         return new CustomerResource($customer);
@@ -58,9 +64,12 @@ class CustomerController extends Controller
      *
      * @param Customer $customer
      * @return CustomerResource
+     * @throws AuthorizationException
      */
     public function show(Customer $customer)
     {
+        $this->authorize('show', $customer);
+
         return new CustomerResource($customer);
     }
 
@@ -70,9 +79,12 @@ class CustomerController extends Controller
      * @param UpdateRequest $request
      * @param Customer $customer
      * @return CustomerResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, Customer $customer)
     {
+        $this->authorize('update', $customer);
+
         $customer = $this->customerRepository->update($customer, $request->all());
 
         return new CustomerResource($customer);
@@ -83,9 +95,12 @@ class CustomerController extends Controller
      *
      * @param Customer $customer
      * @return \Illuminate\Http\Response
+     * @throws AuthorizationException
      */
     public function destroy(Customer $customer)
     {
+        $this->authorize('destroy', $customer);
+
         $this->customerRepository->delete($customer);
 
         return response()->json(null, 204);
